Add HTTP/3 columns
Add an HTTP/3 column to server and CDN/PaaS tables.
In the interest of simplicity, this column reflects
support for the final RFC version of HTTP/3. Since
the IETF are yet to publish this version, all entries
begin with the value of "no".

Over time, we'll populate cells and amend this commit.
Sensible thresholds might include HTTP/3 support merged
into server mainline codebase (or module equivalents)
and CDN/PaaS betas that have public announcements.