docs: Better clarify ECDSA support
Emphasizes lack of support is typically for clients from 2008 or earlier by referring to such clients as legacy and the usual context of this being in an enterprise environment, additionally linking to a CA compatibility page that cites references. 

It should help instill confidence in serving ECDSA for the common modern deployments of websites and reduce concerns in ECDSA adoption when a server is unable to provide dual/hybrid certificate support (Caddy and k8s nginx ingress controller are two popular choices that come to mind).